refactoring chart
